+++ Sources/StreamChatSwiftUI/ChatMessageList/MessageListHelperViews.swift
@@ -180,9 +180,6 @@ extension View {
     public func textColor(for message: ChatMessage) -> Color {
         @Injected(\.colors) var colors
         
-        if message.isDeleted {
-            return Color(colors.textTertiary)
-        }
         if message.isSentByCurrentUser {
             return Color(colors.chatTextOutgoing)
         } else {